Use HTML5 elements to structure the page layout.
Moves the code to a mode consistent layout - both with and without
a navbar - and take advantage of HTML5 constructs to provide semantic
structure to the pages (e.g. header, footer, main, aside, and nav).

At present the only nav element is provided for the site navigation
links in the header. The "navigation" links in the "nav area" to
the left of the page are not marked up using the nav tag at this
time.