Basic support for screen blocks that you want to allow to wrap
The current setup does not make it easy to have wide screen blocks,
since they end up making the display too wide. It is not clear to me
if changes in the HTML5 move (or related CSS changes) have made this
happen, or it always did and the "overflow: auto" rule on these were
never being triggered.

As a simple hack, you can now add a wrap="yes" attribute to screen tags
to allow the text to wrap. This is really meant to be a short-term fix,
but I expect it to remain for a while...