Include the contents of banner.html in every page
The banner.html contents are included using server-side include, so
acts as a way of adding information to each page without having to
re-publish it. It is not ideal, but we are using SSI anyway so it is
simple.

Note that styling is assumed to be included in the banner file (or it
is empty). As the file gets included twice this is a bit of a waste
but let's see how well it works.

There's currently no way to override the banner.html file name from
the publish script, but it would not be hard to do so.