If /*/info/breadcrumbs exists then add a "breadcrumb" trail
A set of links, listing the location of this page within the
hiearachy of the site, in the form

  / site / dir1 / ... / page.html

are created at the top and bottom of the main section of the page,
where each segment (other than page.html) is a link.

The stylesheets need to have something like the following, where the
parent of .breadcrumbs is position: relative.

  position: relative;
}

.breadcrumbs {
  background: #eeeeee;
  font-family: monospace;
  font-size: smaller;
  margin-right: 1em;
  margin-top: 0.5em;
  position: absolute;
  right: 1em;
}

.breadcrumbs-top {
  top: 0.5em;
}

.breadcrumbs-bottom {
  bottom: 0.5em;
}