Do not publish directory if contains DO_NOT_PUBLISH file
A "quick" way of excluding a directory from being published
if it contains a DO_NOT_PUBLISH file, or any parent directory
up to the documentation base. This is to support adding a
"data" directory to the documentation containing images/code/whatever
used to create the documentation, so needs to "live" with the
documentation, but it shouldn't be published.