Test of notebook handling.
This technically works, but something is odd with the
test notebook I'm using (the widths of the putput blocks are
restricted and after about 4 cells everything starts moving
to the right). This could be because of the way I am
breaking up the HTML output, or something else.

I may need to actually work out how to use templates with
nbconvert to do the conversion myself.