Improvements for code[@lang] blocks
We now explicitly check for pygmentize being present before
processing a file containing code blocks with a lang tag, and
error out with a hopefully useful error message.

We need CSS rules for when the code is styled. The options are
to read them in from a single location or to include them in the
file being procesed. For now we go with the latter as it is easier
to set up, and is easier to handle code style changes due to the
pygmentize code changing. However this is still very experimental.