fix: screen handling with lang tag
We now only process the screen contents (to add language styling)
if the lang tag is set (i.e. language != 'none'). This then lets
us disable entity escaping in this case (since the code will have
added tags we want to have handled as HTML) but keep it for the
case when there is no post-processing.

This shold fix it. Definitely. No reason for more commits.