ahelp: ensure repeated title attributes do not create multiple id attributes
The code automatically adds id attributes for headers if the ADESC
or PARA block has a title attribute. Since we only have the title
string we convert it into a hopefully-valid anchor (by replacing
each space characer with _). However, we have ahelp files where the
same title is repeated, leading to invalid output as it would have
multiple id attributes with the same anchor.

So, for these elements we just drop adding the id attribute, and
report it to the screen (it's a bit annoying becuse you get the
same message repeated for each match).