myhtml: better-handle item tags
The way Ihave set up the xi:include handling we often end up
including the "item" tag used to hold the contents. As this is
so common let's just process the contents of these tags, and let
the user know it's done in case something funky is going on (but
who's going to track this ...).