Ensure links to ".../" are replaced by ".../index.html"
This is for CXC pages only. There is at least one place where we
do not do this, which is for the /help/ link, since this appears
to fail if you go to /help/index.html (I assume that however we
plumb in osTicket support we don't handle this variant).

The idea is here to make the pages work better with a local
installation (i.e. the pages are not being served by apache).