news: fix the anchors so that they are unique
We had been using news.html#item-YYYY-MM-DD for the anchors for
the individual items, but this fails miserably when there are
multiple items released on the same day.

So, we use news.html#item-YYYY-MM-DD-<ctr> where ctr is 1 .. N,
with the first entry for a particular day being 1 (so newer items
in the same day have later ctr values). We always add a ctr value,
even if there is only one item, in case we have multiple items
added on the same day but not exactly the same time.