Allow multiple CSS blocks when using pygmentize
There's no need to fall over if there are multiple CSS blocks;
we can just add to the first one, which will hopefully mean
we can over-ride settings if need be. Alternatively we could just
add our own CSS block to the start of the file but let's not
go there just yet.