Skip XInclude when checking for changed files
My use of XInclude files in the sherpa FAQ has shown a problem in the
existing code. It looks like we need to copy the XInclude file to the
repository/storage area. However, this is "tricky" and as we have not
needed it now we just skip the XInclude processing when running this
check. This will likely come and bite us at some point, but not today.

I also don't know why this has not been a problem before, as we use
XInclude statements elsewhere. Maybe never for a "special" case
(like the faq or dict pages) which involve key checks?