find and store xinclude files in a page
We also try to use the xinclude date in the check to see when the
information was last updated (just because the xinclude file has
changed does not mean the page has changed, but it's simplest to
assume so).

I have forgotten so much about thios, including why we only appear to
store information in the store/punlished area if we have
dependencies. In some ways it makes sense as we only need these files
if we have the dependencies.