Explicitly expand the XInclude contents for published files
I managed to cause a tiz by using a XInclude file in something that
triggers the dependency tracking, as the contents of the XInclude file
were not copied over. This lead to some changes. I realised that if we
copied over the file *after* expanding the XInclude then we could
avoid this complication. It turns out that some of the extra code to
track the XInclude files is technically useful (the tracking of
whether to re-publish the file), so it wasn't all wasted.